+++ cachelib/__init__.py
@@ -4,6 +4,6 @@ from .memory import MemoryCache
 
 
 __all__ = ["MemoryCache"]
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __author__ = "Deetjepateeteke <https://github.com/Deetjepateeteke>"
 __license__ = "MIT"